# docs/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/endpoints": {
			"post": {
				"description": "Creates a new proxy endpoint. A unique slug is auto-generated.",
				"operationId": "EndpointsController_create",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/Function" }
						}
					}
				},
				"responses": {
					"201": { "description": "Endpoint created successfully" },
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Create endpoint",
				"tags": ["Endpoints"]
			},
			"get": {
				"description": "Returns all proxy endpoints for the authenticated user.",
				"operationId": "EndpointsController_findAll",
				"parameters": [],
				"responses": {
					"200": { "description": "List of endpoints" },
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "List endpoints",
				"tags": ["Endpoints"]
			}
		},
		"/api/v1/endpoints/{id}": {
			"get": {
				"description": "Returns a single endpoint by ID. User must own the endpoint.",
				"operationId": "EndpointsController_findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Endpoint details" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"404": {
						"description": "Not Found - Resource does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Get endpoint by ID",
				"tags": ["Endpoints"]
			},
			"patch": {
				"description": "Updates an existing endpoint. Only provided fields are updated.",
				"operationId": "EndpointsController_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/Function" }
						}
					}
				},
				"responses": {
					"200": { "description": "Endpoint updated successfully" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"404": {
						"description": "Not Found - Resource does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Update endpoint",
				"tags": ["Endpoints"]
			},
			"delete": {
				"description": "Permanently deletes an endpoint and its associated data.",
				"operationId": "EndpointsController_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Endpoint deleted successfully" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"404": {
						"description": "Not Found - Resource does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Delete endpoint",
				"tags": ["Endpoints"]
			}
		},
		"/api/v1/logs/endpoint/{endpointId}": {
			"get": {
				"description": "Returns request logs for an endpoint with optional filtering.",
				"operationId": "LogsController_findByEndpoint",
				"parameters": [
					{
						"name": "endpointId",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					},
					{
						"name": "limit",
						"required": false,
						"in": "query",
						"description": "Max results (default: 50)",
						"schema": { "type": "string" }
					},
					{
						"name": "offset",
						"required": false,
						"in": "query",
						"description": "Pagination offset",
						"schema": { "type": "string" }
					},
					{
						"name": "method",
						"required": false,
						"in": "query",
						"description": "Filter by HTTP method",
						"schema": { "type": "string" }
					},
					{
						"name": "status",
						"required": false,
						"in": "query",
						"description": "Filter by response status",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "List of request logs" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "List logs by endpoint",
				"tags": ["Logs"]
			}
		},
		"/api/v1/logs/{id}": {
			"get": {
				"description": "Returns a single request log by ID.",
				"operationId": "LogsController_findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Log UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Request log details" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"404": {
						"description": "Not Found - Resource does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Get log by ID",
				"tags": ["Logs"]
			}
		},
...
		"/api/v1/notifications/alert-rules/endpoint/{endpointId}": {
			"get": {
				"description": "Returns all alert rules for a given endpoint.",
				"operationId": "NotificationsController_getAlertRules",
				"parameters": [
					{
						"name": "endpointId",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "List of alert rules" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "List alert rules for endpoint",
				"tags": ["Notifications"]
			}
		},
...
	},
	"info": {
		"title": "Proxy Server API",
		"description": "API documentation for the Proxy Server backend",
		"version": "1.0",
		"contact": { "name": "Proxy Server", "url": "", "email": "" },
		"license": { "name": "ISC", "url": "" }
	},
...
	"servers": [
		{ "url": "http://localhost:3000", "description": "development" },
		{ "url": "http://lvh.me:3000", "description": "production" }
	],
	"components": {
		"securitySchemes": {
			"Bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter JWT access token from /auth/sign-in, /auth/verify-email, or /auth/refresh",
				"in": "header",
				"name": "Authorization",
				"type": "http"
			}
		},
		"schemas": {
...
			"Function": { "type": "object", "properties": {} },
			"ErrorResponseSchema": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "number",
						"description": "HTTP status code",
						"example": 400,
						"minimum": 400,
						"maximum": 599
					},
					"error": {
						"type": "string",
						"description": "Error type or exception name",
						"example": "BadRequestException"
					},
					"message": {
						"description": "Human-readable error message(s)",
						"oneOf": [
							{ "type": "string", "example": "Validation failed" },
							{
								"type": "array",
								"items": { "type": "string" },
								"example": [
									"email must be an email",
									"password must be longer than 8 characters"
								]
							}
						]
					},
					"path": {
						"type": "string",
						"description": "Request path that triggered the error",
						"example": "/api/v1/auth/register"
					},
					"timestamp": {
						"type": "string",
						"description": "ISO 8601 timestamp when the error occurred",
						"example": "2025-03-18T12:00:00.000Z"
					}
				},
				"required": ["statusCode", "error", "message"]
			}
		}
	},
	"security": [{ "Bearer": [] }]
}